Final Merge of v1 smart account ( Testnet )

+++ src/PushLocker/PushLocker.sol
@@ -9,18 +9,10 @@ import {IERC20} from "@openzeppelin/contracts/token/ERC20/IERC20.sol";
 import {SafeERC20} from "@openzeppelin/contracts/token/ERC20/utils/SafeERC20.sol";
 import {ReentrancyGuardUpgradeable} from "@openzeppelin/contracts-upgradeable/utils/ReentrancyGuardUpgradeable.sol";
 
-contract PushLocker is
-    Initializable,
-    UUPSUpgradeable,
-    AccessControlUpgradeable,
-    ReentrancyGuardUpgradeable
-{
+contract PushLocker is Initializable, UUPSUpgradeable, AccessControlUpgradeable, ReentrancyGuardUpgradeable {
     using SafeERC20 for IERC20;
-    event FundsAdded(
-        address indexed user,
-        uint256 usdtAmount,
-        bytes32 transactionHash
-    );
+
+    event FundsAdded(address indexed user, uint256 usdtAmount, bytes32 transactionHash);
     event TokenRecovered(address indexed admin, uint256 amount);
 
     address public WETH;
@@ -30,13 +22,10 @@ contract PushLocker is
 
     uint24 constant POOL_FEE = 500; // 0.05%
 
-    function initialize(
-        address _admin,
-        address _weth,
-        address _usdt,
-        address _router,
-        address _priceFeed
-    ) external initializer {
+    function initialize(address _admin, address _weth, address _usdt, address _router, address _priceFeed)
+        external
+        initializer
+    {
         __ReentrancyGuard_init();
         _grantRole(DEFAULT_ADMIN_ROLE, _admin);
 
@@ -46,50 +35,42 @@ contract PushLocker is
         ethUsdPriceFeed = AggregatorV3Interface(_priceFeed);
     }
 
-    function _authorizeUpgrade(
-        address newImplementation
-    ) internal override onlyRole(DEFAULT_ADMIN_ROLE) {}
+    function _authorizeUpgrade(address newImplementation) internal override onlyRole(DEFAULT_ADMIN_ROLE) {}
 
     function addFunds(bytes32 _transactionHash) external payable nonReentrant {
         require(msg.value > 0, "No ETH sent");
 
         // Wrap ETH to WETH
         IWETH(WETH).deposit{value: msg.value}();
-        uint WethBalance = IERC20(WETH).balanceOf(address(this));
+        uint256 WethBalance = IERC20(WETH).balanceOf(address(this));
         IWETH(WETH).approve(UNISWAP_ROUTER, WethBalance);
 
         // Get current ETH/USD price from Chainlink
-        (, int256 price, , , ) = ethUsdPriceFeed.latestRoundData(); // price is 8 decimals
+        (, int256 price,,,) = ethUsdPriceFeed.latestRoundData(); // price is 8 decimals
         require(price > 0, "Invalid oracle price");
 
-        uint256 ethInUsd = (uint256(price) *WethBalance) / 1e8;
+        uint256 ethInUsd = (uint256(price) * WethBalance) / 1e8;
 
         // Expect similar USDT amount (1:1 with USD), allow 0.5% slippage
         uint256 minOut = (ethInUsd * 995) / 1000;
 
-        ISwapRouter.ExactInputSingleParams memory params = ISwapRouter
-            .ExactInputSingleParams({
-                tokenIn: WETH,
-                tokenOut: USDT,
-                fee: POOL_FEE,
-                recipient: address(this),
-                 deadline: block.timestamp, //not for sepolia
-                amountIn: WethBalance,
-                amountOutMinimum: minOut / 1e12, // Adjust to USDT decimals (6) && not for sepolia
-                sqrtPriceLimitX96: 0
-            });
+        ISwapRouter.ExactInputSingleParams memory params = ISwapRouter.ExactInputSingleParams({
+            tokenIn: WETH,
+            tokenOut: USDT,
+            fee: POOL_FEE,
+            recipient: address(this),
+            deadline: block.timestamp, //not for sepolia
+            amountIn: WethBalance,
+            amountOutMinimum: minOut / 1e12, // Adjust to USDT decimals (6) && not for sepolia
+            sqrtPriceLimitX96: 0
+        });
 
-        uint256 usdtReceived = ISwapRouter(UNISWAP_ROUTER).exactInputSingle(
-            params
-        );
+        uint256 usdtReceived = ISwapRouter(UNISWAP_ROUTER).exactInputSingle(params);
 
         emit FundsAdded(msg.sender, usdtReceived, _transactionHash);
     }
 
-    function recoverToken(
-        address _recipient,
-        uint256 amount
-    ) external onlyRole(DEFAULT_ADMIN_ROLE) {
+    function recoverToken(address _recipient, uint256 amount) external onlyRole(DEFAULT_ADMIN_ROLE) {
         IERC20(USDT).safeTransfer(_recipient, amount);
 
         emit TokenRecovered(_recipient, amount);
